Progwards unit 9

+++ src/ru/progwards/java1/lessons/bigints/ByteInteger.java
@@ -1,12 +1,10 @@
 package ru.progwards.java1.lessons.bigints;
 
-import java.math.BigInteger;
-
 public class ByteInteger extends AbsInteger {
     byte num;
 
     public ByteInteger(byte num) {
-        super(BigInteger.valueOf(num));
+        super(num);
         this.num = num;
     }
