+++ src/ru/progwards/java1/lessons/bigints/IntInteger.java
@@ -1,12 +1,10 @@
 package ru.progwards.java1.lessons.bigints;
 
-import java.math.BigInteger;
-
 public class IntInteger extends AbsInteger{
     int num;
 
     public IntInteger(int num) {
-        super(BigInteger.valueOf(num));
+        super(num);
         this.num = num;
     }
