+++ src/ru/progwards/java1/lessons/bigints/ShortInteger.java
@@ -1,12 +1,10 @@
 package ru.progwards.java1.lessons.bigints;
 
-import java.math.BigInteger;
-
 public class ShortInteger extends AbsInteger {
     short num;
 
     public ShortInteger(short num) {
-        super(BigInteger.valueOf(num));
+        super(num);
         this.num = num;
     }
